[Glimmer 2] Implement mut and readonly helpers
Also, implement support for the [INVOKE] symbol in closure actions.

This also enables (partial) "reference pooling" on components – when
doing rending a `{{foo}}` in the `{{foo-bar}}` layout, it first checks
if `{{foo}}` is an arg passed from the outside (`{{foo-bar foo=...}}`),
and if so, it tries to reuse the same reference instead of making a new
one. This should reduce the total number of references created and also
reduce the number of `value()` computation as the same reference can be
reused across multiple curlies.

This also allows Glimmer to better optimize things – for example, if the
component is invoked with literals (`{{foo-bar foo=true}}`) and `foo` is
then used in a conditional in its layout (`{{#if foo}}...`), Glimmer
will be now able to see that `foo` is a const reference and optimize out
the updating step. This works recursively as well – if `foo-bar` then
passes the same property down (`{{bar-baz bar=foo}}`), we now propagate
the same reference all the way down.